refactor : clubBoard로 디렉토리 변경
