+++ Server/main07/src/main/java/com/codestates/main07/clubBoard/comment/dto/MemoryBoardCommentCreateDto.java
@@ -1,0 +1,9 @@
+package com.codestates.main07.clubBoard.comment.dto;
+
+import lombok.Getter;
+
+@Getter
+public class MemoryBoardCommentCreateDto {
+    private long memberId;
+    private String content;
+}
