+++ Server/main07/src/main/java/com/codestates/main07/clubBoard/comment/dto/MemoryBoardCommentUpdateDto.java
@@ -1,0 +1,9 @@
+package com.codestates.main07.clubBoard.comment.dto;
+
+import lombok.Getter;
+
+@Getter
+public class MemoryBoardCommentUpdateDto {
+    private long memberId;
+    private String content;
+}
