+++ Server/main07/src/main/java/com/codestates/main07/clubBoard/comment/repository/MemoryBoardCommentRepository.java
@@ -1,0 +1,9 @@
+package com.codestates.main07.clubBoard.comment.repository;
+
+import com.codestates.main07.clubBoard.comment.entity.MemoryBoardComment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemoryBoardCommentRepository extends JpaRepository<MemoryBoardComment, Long> {
+}
