+++ Server/main07/src/main/java/com/codestates/main07/clubBoard/response/SuccessDto.java
@@ -1,0 +1,14 @@
+package com.codestates.main07.clubBoard.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SuccessDto {
+    private boolean success;
+
+    public SuccessDto(boolean success) {
+        this.success = success;
+    }
+}
