--- Server/main07/src/main/java/com/codestates/main07/memoryBoard/board/repository/MemoryBoardRepository.java
@@ -1,9 +1,0 @@
-package com.codestates.main07.memoryBoard.board.repository;
-
-import com.codestates.main07.memoryBoard.board.entity.MemoryBoard;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemoryBoardRepository extends JpaRepository<MemoryBoard, Long> {
-}
