--- Server/main07/src/main/java/com/codestates/main07/memoryBoard/comment/dto/MemoryBoardCommentUpdateDto.java
@@ -1,9 +1,0 @@
-package com.codestates.main07.memoryBoard.comment.dto;
-
-import lombok.Getter;
-
-@Getter
-public class MemoryBoardCommentUpdateDto {
-    private long memberId;
-    private String content;
-}
