--- Server/main07/src/main/java/com/codestates/main07/memoryBoard/response/PageInfo.java
@@ -1,13 +1,0 @@
-package com.codestates.main07.memoryBoard.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class PageInfo {
-    private int page;
-    private int size;
-    private long totalElements;
-    private int totalPages;
-}
