--- Server/main07/src/main/java/com/codestates/main07/memoryBoard/response/SuccessDto.java
@@ -1,14 +1,0 @@
-package com.codestates.main07.memoryBoard.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SuccessDto {
-    private boolean success;
-
-    public SuccessDto(boolean success) {
-        this.success = success;
-    }
-}
